% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{buildMetaLevelDataset}
\alias{buildMetaLevelDataset}
\title{Builds the meta-level data set}
\usage{
buildMetaLevelDataset(timeseries, natural = TRUE, boxcox = TRUE)
}
\arguments{
\item{timeseries}{A list of time series}

\item{natural}{Optional parameter: A flag indicating wheter only natural frequencies (e.g., daily, hourly, ...) or all found frequencies shall be considered.}

\item{boxcox}{Optional parameter: A flag indicating if the Box-Cox transofrmation should be performed. It is not recommend to disable the transformation. TRUE by default.}
}
\value{
The accuracies and characteristics of the time series.
}
\description{
Builds the meta-level data set. That is, calculating the forecasting error and time series characteritics for each time series.
}
