% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_periods.R
\name{calcClustersForPeriods}
\alias{calcClustersForPeriods}
\title{Calculates the Clusters for Periods}
\usage{
calcClustersForPeriods(timeseries, frequency, doAnomDet, replace.zeros,
  debug = FALSE)
}
\arguments{
\item{timeseries}{The time series to split}

\item{frequency}{The determined frequency of the time series}

\item{doAnomDet}{Boolean whether anomaly detection shall be used for clustering}

\item{replace.zeros}{If TRUE, all zeros will be replaced by the mean of the non-zero neighbors}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}
}
\value{
The cluster label for each period
}
\description{
Calculates the clusters for the periods of the time series regarding mean, variance and range
}
\author{
Marwin Zuefle, Andre Bauer
}
