% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{calcFrequencyPeriodogram}
\alias{calcFrequencyPeriodogram}
\title{Calcuates the Frequency of the Time Series}
\usage{
calcFrequencyPeriodogram(timeValuePair, asInteger = TRUE, difFactor = 0.5,
  tolerance = 0.05, maxIters = 10, ithBest = 1, PGramTvp = NULL)
}
\arguments{
\item{timeValuePair}{Timestamps with raw values (not as time series object)}

\item{asInteger}{Optional parameter: Boolean indicating whether frequerncy needs to be an integer. TRUE by default}

\item{difFactor}{Optional parameter: A factor to determine whether a possible frequency is accepted. 0.5 by default}

\item{tolerance}{Optional parameter: A tolerance factor for matching estimated frequencies to listed frequencies. 0.05 by default}

\item{maxIters}{Optional parameter: Number of maximal rounds to estimate the frequency. 10 by default.}

\item{ithBest}{Optional parameter: i-th most likely frequency is tested. 1 by default.}

\item{PGramTvp}{Optional parameter: An already created periodogram. Null by default}
}
\value{
The frequency that appears in the periodogram and the list of most frequent frequencys, the created perodogram and the last iteration
}
\description{
Calculates the frequency for the time series
}
