% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{calculateAccuracies}
\alias{calculateAccuracies}
\title{Calculates the forecasting error of each method.}
\usage{
calculateAccuracies(data, label)
}
\arguments{
\item{data}{The features of the time series.}

\item{label}{The detrended time series.}
}
\value{
The auto-correlation of the errors of the fit
}
\description{
Quantifies how well each method has performed on the given time series.
}
