% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{calculateCharacteristics}
\alias{calculateCharacteristics}
\title{Calculates time series characteristics.}
\usage{
calculateCharacteristics(tvp.stl, freq)
}
\arguments{
\item{tvp.stl}{The STL decomposistion of the time series}

\item{freq}{The frequency of the time series}
}
\value{
The time series characteritics
}
\description{
Calculates the time series characteristics of the time series.
}
