% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{check.trend}
\alias{check.trend}
\title{Estimating the boosting method}
\usage{
\method{check}{trend}(stl.decomp, lower.percentile = 0.05,
  upper.percentile = 0.95, threshold = 0.33)
}
\arguments{
\item{stl.decomp}{The STL decomposition of a time series}

\item{lower.percentile}{Optional parameter: The lower percentile for the IPR. 0.05 by default}

\item{upper.percentile}{Optional parameter: The upper percentile for the IPR. 0.95 by default}

\item{threshold}{Optional parameter: Threshold what the propotion of trend component has to exceed to be trendy. 0.33 by default}
}
\value{
True if the time series has a significant trend
}
\description{
Checks if the time series has a significant trend
}
