% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{computeMASEsameValue}
\alias{computeMASEsameValue}
\title{Compute MASE for same value}
\usage{
computeMASEsameValue(forecast, train, test, plot)
}
\arguments{
\item{forecast}{The forecasted values.}

\item{train}{The 'historical' data used for forecasting.}

\item{test}{The 'future' data used for finding MASE.}

\item{plot}{Boolean indicating whether the forecast should be plotted.}
}
\value{
MASE between forecast and real data
}
\description{
MASE for a naive forecast taking the last observation for the whole forecast
}
