% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender.R
\name{consultrecommender}
\alias{consultrecommender}
\title{Retrieves the best method for a given time series.}
\usage{
consultrecommender(tvp, tvp.stl = NULL, model, natural = TRUE, boxcox = TRUE)
}
\arguments{
\item{tvp}{A time series. Will be ignored if tvp.stl is set.}

\item{tvp.stl}{Optional parameter: A stl decomposition of the time series.}

\item{model}{A object of the class telescope.recommendation containing the hybrind random forest model.}

\item{natural}{Optional parameter: A flag indicating wheter only natural frequencies (e.g., daily, hourly, ...) or all found frequencies shall be considered.}

\item{boxcox}{Optional parameter: A flag indicating if the Box-Cox transofrmation should be performed. It is not recommend to disable the transformation. TRUE by default.}
}
\value{
The best method as string.
}
\description{
Consults the recommendation system to retrieve the best method for a given time series.
}
