% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{doANN}
\alias{doANN}
\title{Apply ANN}
\usage{
doANN(myts, rep = 20)
}
\arguments{
\item{rep}{The amount of repeats}

\item{ts}{The timeseries.}
}
\value{
The ANN Forecast
}
\description{
Performs the ANN forecast of the timeseries.
}
