% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{doXGB.train}
\alias{doXGB.train}
\title{XGBoost Model Training}
\usage{
doXGB.train(myts, cov, booster, verbose)
}
\arguments{
\item{myts}{The training part of the time series}

\item{cov}{The covariates for the training}

\item{booster}{Select a boosting method: gblinear or gbtree}

\item{verbose}{Set to 1 to print information of xgboost performance; 0 for no prints}
}
\value{
model The XGBoost model
}
\description{
Trains XGBoost regarding the time series and its covariates
}
\author{
Marwin Zuefle, Andre Bauer
}
