% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{durbinwatson}
\alias{durbinwatson}
\title{Checks the auto-correlation of the fitted errors.}
\usage{
durbinwatson(e)
}
\arguments{
\item{e}{The error of a fit.}
}
\value{
The auto-correlation of the errors of the fit
}
\description{
Quantifies how well the time series can be approximated by a sinus wave.
}
