% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{extract.ts}
\alias{extract.ts}
\title{Extract time series information}
\usage{
\method{extract}{ts}(tvp, debug = FALSE)
}
\arguments{
\item{tvp}{The time value pair: either vector of raw values or n-by-2 matrix (raw values in second column), or time series}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}
}
\value{
the time value pair as vector, the frequency of the data, use.sec.freq and if the data was no time series, the last iteration and the periodigram of the frequncy estimation is also returned
}
\description{
Extract the requird information (e.g., frequency, values) of the time series
}
