% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{fitSin}
\alias{fitSin}
\title{Fitting a sinus.}
\usage{
fitSin(y)
}
\arguments{
\item{y}{A vector containg the time series.}
}
\value{
The errors of the fit.
}
\description{
Fits a sinus on the data and returns the fitted errors.
}
