% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_models.R
\name{fittingModels}
\alias{fittingModels}
\title{Fitting the Model of the Trend}
\usage{
fittingModels(stl, frequency, difFactor = 1.5, debug = FALSE)
}
\arguments{
\item{frequency}{The frequency of the time-value pair}

\item{difFactor}{Optional parameter: The factor how much exp needs to be better than linear. 1.5 by default}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}

\item{tvp}{The time-value pair}
}
\value{
The model, either linear or exp and if the estimation is riskys
}
\description{
Checks the model of the trend: either linear or exponential
}
\author{
Marwin Zuefle, Andre Bauer
}
