% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{forecast.season}
\alias{forecast.season}
\title{Forecasting Season}
\usage{
\method{forecast}{season}(total.length, season, frequency, hist.length)
}
\arguments{
\item{total.length}{history + horizon}

\item{season}{The seasonal pattern from stl}

\item{frequency}{The frequency of the timeseries}

\item{hist.length}{The length of the history}
}
\value{
The seasonal pattern for the forecast horizon
}
\description{
Forecasts the season part of the time series
}
