% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluster_periods.R
\name{forecastClusters}
\alias{forecastClusters}
\title{Forecast the Clusters}
\usage{
forecastClusters(clusters, frequency, timeseries, reps, horizon,
  debug = FALSE)
}
\arguments{
\item{clusters}{The clusters found for the periods}

\item{timeseries}{The time series}

\item{reps}{The amount of repeats for ANN}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}

\item{freuqency}{The determined frequency}
}
\value{
All cluster labels (history and forecast)
}
\description{
Forecasts the clusters for the forecast periods
}
