% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frequency.R
\name{guessFrequencyPeriodogram}
\alias{guessFrequencyPeriodogram}
\title{Guess the Frequency of the Time Series}
\usage{
guessFrequencyPeriodogram(
  timeValuePair,
  asInteger = TRUE,
  difFactor = 0.5,
  ithBest = 1,
  spans = NULL,
  debug = FALSE,
  PGramTvp = NULL
)
}
\arguments{
\item{timeValuePair}{Timestamps with raw values (not as time series object)}

\item{asInteger}{Optional parameter: Boolean indicating whether frequerncy needs to be an integer}

\item{difFactor}{Optional parameter: A factor to determine whether a possible frequency is accepted}

\item{ithBest}{Optional parameter: i-th most likely frequency is tested. 1 by default.}

\item{spans}{Optional parameter: vector of odd integers giving the widths of modified Daniell smoothers to be used to smooth the periodogram. Null by default}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}

\item{PGramTvp}{Optional parameter: An already created periodogram. Null by default}
}
\value{
The found frequency and the created periodogram
}
\description{
Estimates the frequency of the time series based on a periodogram
}
\author{
Marwin Zuefle, Andre Bauer
}
