% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{has.highRemainder}
\alias{has.highRemainder}
\title{Checking Remainder}
\usage{
has.highRemainder(tvp, stlRemainder, use.log, sig.dif.factor)
}
\arguments{
\item{tvp}{tvp The time value pair as vector}

\item{stlRemainder}{The remainder part of the deccomposition of the tvp}

\item{use.log}{A flag if log was used for tvp}

\item{sig.dif.factor}{The threshold that is to exceed for having a high remainder}
}
\value{
If the the time series has a high remainder compared to the threshold
}
\description{
Checks if the time series has a significant high remainder
}
