% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xgb.R
\name{IPR}
\alias{IPR}
\title{Inter Percentile Range}
\usage{
IPR(x, lower.percentile, upper.percentile)
}
\arguments{
\item{x}{The vector to compute interpercentile range of}

\item{lower.percentile}{The lower percentile}

\item{upper.percentile}{The upper percentile}
}
\value{
The range between the lower and upper percentile
}
\description{
computes the interpercentile range of vector x from percentile lower.percentile to percentile upper.percentile
}
