% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{mape}
\alias{mape}
\title{Calculates forecasting error.}
\usage{
mape(forecast, test)
}
\arguments{
\item{forecast}{The forecast of the detrended time series.}

\item{test}{The acutual detrended time series.}
}
\value{
The mape measure.
}
\description{
Calculates the error of the forecast regarding the acutal values.
}
