% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildmetaleveldataset.R
\name{nnetarprediction}
\alias{nnetarprediction}
\title{Trains a nnetar model.}
\usage{
nnetarprediction(train, label, test)
}
\arguments{
\item{train}{A matrix containing the seasonal component and fourier terms for training.}

\item{label}{A vector containing the detrended time series.}

\item{test}{A matrix containing the seasonal component and fourier terms for the prediction.}
}
\value{
The future detrended time series.
}
\description{
Trains a nnetar model.
}
