% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBubble.R
\name{plotBubble}
\alias{plotBubble}
\title{Plot multivariate categorical data.}
\usage{
plotBubble(x, bg = "grey")
}
\arguments{
\item{x}{A matrix or data.frame; the data consisting of categorical
variables.}

\item{bg}{If specified, the symbols are filled with colour(s), the
vector \code{bg} is recycled to the number of observations.  The
default is to fill the symbols with grey color.}
}
\value{
\code{NULL}
}
\description{
Plots of the level combinations of pairs of variables
are created where the size of the circle indicating a level
combination is proportional to the frequency of the level
combination.
}
\examples{
with(chickwts, plotBubble(data.frame(cut(weight, 100 * 1:5), feed)))
}
