% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotScatter.R
\name{plotScatter}
\alias{plotScatter}
\title{Pairwise scatter plots of the data.}
\usage{
plotScatter(x, z, label = "", trim = 0)
}
\arguments{
\item{x}{A matrix or data.frame; the data consisting of metric variables.}

\item{z}{A vector; indicating the color to use for the observations.}

\item{label}{A character string; the text to include in the axes labels.}

\item{trim}{A scalar numeric in \eqn{[0, 0.5)}; trimming to use for quantiles to determine axes.}
}
\value{
\code{NULL}
}
\description{
Scatter plots of the observations are plotted by
selecting pairs of dimensions, potentially colored by a known
classification.
}
\examples{
plotScatter(iris[, 1:4], iris$Species, label = "dim")
}
