% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope.R
\name{plotting}
\alias{plotting}
\title{Plots the forecast}
\usage{
plotting(output)
}
\arguments{
\item{output}{contains the forecast, the fitted model, and the time series}
}
\description{
Plots the modeling of the history and the forecast
}
