% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope.R
\name{preprocessing}
\alias{preprocessing}
\title{Perform the Forecast}
\usage{
preprocessing(tvp, natural = TRUE, boxcox = TRUE)
}
\arguments{
\item{tvp}{The time value pair: a time series}

\item{natural}{Optional parameter: A flag indicating wheter only natural frequencies (e.g., daily, hourly, ...) or all found frequencies shall be considered.}

\item{boxcox}{Optional parameter: A flag indicating if the Box-Cox transofrmation should be performed. It is not recommend to disable the transformation. TRUE by default.}
}
\value{
The adjusted time series, the dominant frequencies, and adjustment parameters
}
\description{
Preprocesses the time series and retrieves dominant frequencies
}
