% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier.R
\name{removeAnomalies}
\alias{removeAnomalies}
\title{Remove the Anomalies}
\usage{
removeAnomalies(rawValues, frequency, replace.zeros)
}
\arguments{
\item{rawValues}{The raw values as vector of the time series without timestamp}

\item{frequency}{The frequency of the time series}

\item{replace.zeros}{If TRUE, all zeros will be replaced by the mean of the non-zero neighbors}
}
\value{
The vector of observations without anomalies (anomalies replaced by mean of normal values around)
}
\description{
Removes the anomalies of the time series
}
\author{
Marwin Zuefle, Andre Bauer
}
