% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleK.R
\name{sampleK_e0}
\alias{sampleK_e0}
\title{Sample K conditional on e0 (fixed or random, but not depending on K).}
\usage{
sampleK_e0(Kp_j, Kmax, log_pK, log_p_e0, e0, N)
}
\arguments{
\item{Kp_j}{A number; indicating the current value of \eqn{K_+}.}

\item{Kmax}{A number; indicating the maximum value of \eqn{K}, for which the conditional posterior is evaluated.}

\item{log_pK}{A function; evaluating the prior of \eqn{K}.}

\item{log_p_e0}{A function; evaluating the log prior of \eqn{e_0}.}

\item{e0}{A number; indicating the value of \eqn{e_0}.}

\item{N}{A number; indicating the number of observations.}
}
\value{
A number indicating the new value of \eqn{K}.
}
\description{
This sampling step only relies on the current
partition and is independent of the current component-specific
parameters, see Frühwirth-Schnatter et al (2021).
}
