% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{save.csv}
\alias{save.csv}
\title{Save as CSV}
\usage{
save.csv(values, name, path = "")
}
\arguments{
\item{values}{The values to be stored in the csv file}

\item{name}{The name of the csv file to be created}

\item{csv.path}{Optional parameter: The path for the saved csv-file. The current workspace by default.}
}
\description{
Saves the vales to a csv-file
}
\author{
Marwin Zuefle, Andre Bauer
}
