% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseData.R
\name{STChooseData}
\alias{STChooseData}
\title{STChooseData}
\usage{
STChooseData(x, gene)
}
\arguments{
\item{x}{an ST data object such as returned by STLoad}

\item{gene}{gene whose expression will be used to provide
visual indication of where spots of interest are.}
}
\description{
Select region of interest in ST spot matrix for display and analysis
}
