% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stload.R
\name{STLoad}
\alias{STLoad}
\title{STLoad}
\usage{
STLoad(f, species = c("hs", "mm"))
}
\arguments{
\item{f}{path to ST data file}

\item{species}{either "hs" or "mm", used for gene annotation}
}
\description{
Load ST data file, extract coordinates, and annotate genes
}
