% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stload.R
\name{STPlotGene}
\alias{STPlotGene}
\title{STPlotGene}
\usage{
STPlotGene(x, genes, size = 1)
}
\arguments{
\item{x}{An ST data object such as returned from STLoad}

\item{gene}{gene identifier. Identifier type is identified
automatically.}
}
\description{
Lookup gene data and create 2D heatmap of expression
}
