% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chooseData.R
\name{STSubset}
\alias{STSubset}
\title{STSubset}
\usage{
STSubset(d, x, y)
}
\arguments{
\item{d}{an ST data object such as returned by STLoad}

\item{x}{Vector of x coords for points to retain}

\item{y}{Vector of x coords for points to retain}
}
\description{
Subset an ST data object base on x,y pairs
}
