% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope.R
\name{telescope.forecast}
\alias{telescope.forecast}
\title{Perform the Forecast}
\usage{
telescope.forecast(tvp, horizon, repsANN = 20, doAnomDet = TRUE,
  replace.zeros = TRUE, use.indicators = TRUE, save_fc = FALSE,
  csv.path = "", csv.name = "Telescope", debug = FALSE)
}
\arguments{
\item{tvp}{The time value pair: either vector of raw values or n-by-2 matrix (raw values in second column), or time series}

\item{horizon}{The number of values that should be forecast}

\item{repsANN}{Optional parameter: The amount of repeats for ANN. 20 by default.}

\item{doAnomDet}{Optional parameter: Boolean whether anomaly detection shall be used for clustering. TRUE by default}

\item{replace.zeros}{Optional parameter: If TRUE, all zeros will be replaced by the mean of the non-zero neighbors. TRUE by default}

\item{use.indicators}{Optional parameter: If TRUE, additional information (e.g. a flag wheter there is a high remainder) will be returned. TRUE by default}

\item{save_fc}{Optional parameter: Boolean wheter the forecast shall be saved as csv. FALSE by default}

\item{csv.path}{Optional parameter: The path for the saved csv-file. The current workspace by default.}

\item{csv.name}{Optional parameter: The name of the saved csvfile. Telescope by default.}

\item{debug}{Optional parameter: If TRUE, debugging information will be displayed. FALSE by default}
}
\value{
The forecast
}
\description{
Forecasts a given univariate time series in a hybrid manner and based on time series decomposition
}
\examples{
telescope.forecast(AirPassengers, horizon=10)
}
\author{
Marwin Zuefle, Andre Bauer
}
