% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender.R
\name{telescope.trainrecommender}
\alias{telescope.trainrecommender}
\title{Trains the recommendation system}
\usage{
telescope.trainrecommender(timeseries, natural = TRUE, boxcox = TRUE)
}
\arguments{
\item{timeseries}{A list of time series}

\item{natural}{Optional parameter: A flag indicating wheter only natural frequencies (e.g., daily, hourly, ...) or all found frequencies shall be considered.}

\item{boxcox}{Optional parameter: A flag indicating if the Box-Cox transofrmation should be performed. It is not recommend to disable the transformation. TRUE by default.}
}
\value{
An object of the class telescope.recommendation
}
\description{
Trains a recommendations for chossing the best-suited machine learning approach for a given time series. Note that non-seasonal time series are ignored.
}
\examples{
telescope.trainrecommender(list(AirPassengers, forecast::taylor))
}
