% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting_models.R
\name{testTrend}
\alias{testTrend}
\title{Test the Signifcant of Trend}
\usage{
testTrend(tvp, frequency)
}
\arguments{
\item{tvp}{The time-value pair (not in log)}

\item{frequency}{The frequency of the time-value pair}
}
\value{
If time series has a significant trend
}
\description{
Checks if the time series has a significant trend
}
