% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recommender.R
\name{trainingrecommender}
\alias{trainingrecommender}
\title{Trains the recommendation system.}
\usage{
trainingrecommender(timeseries.chars, accuracies)
}
\arguments{
\item{timeseries.chars}{A mxn matrix with n time series characteristics of m time series. More precisely, each row reflects the characteristics of a time series.}

\item{accuracies}{A mxn matrix with accuries measures of n machine learning methods based on m time series.}
}
\value{
An object of the class telescope.recommendation.
}
\description{
Trains a recommendations for chossing the best-suited machine learning approach for a given time series.
}
