% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/telescope_Utils.R
\name{forecast.season}
\alias{forecast.season}
\title{Forecasting Season}
\usage{
\method{forecast}{season}(tvp, stlTrain, horizon)
}
\arguments{
\item{tvp}{The time value pair as vector}

\item{stlTrain}{The decomposition of tvp}

\item{horizon}{The forecast horizon}
}
\value{
The seasonal pattern for the forecast horizon
}
\description{
Forecasts the season part of the time series
}
